# backend/generated/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@ankiai.com
    name: AnkiAI Support
  description: API for generating Anki flashcards with text and media content
  title: AnkiAI API
  version: 1.0.0
servers:
- description: Development server
  url: http://localhost:8000/api
security:
- BearerAuth: []
paths:
  /generateText:
    post:
      description: Generate vocabulary flashcards with audio from a text prompt
      operationId: generateText
      requestBody:
        content:
          application/json:
            example:
              prompt: HSK2向けに『替・觉得・认为』。各語義ごとに短義＋例文を。
              targetLang: zh-CN
              withTTS: true
              maxTerms: 10
              maxMeaningsPerTerm: 3
            schema:
              $ref: "#/components/schemas/GenerateTextRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
          description: Successfully generated flashcard content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request - invalid input parameters
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error
      summary: Generate flashcard content from text prompt
      tags:
      - content-generation
  /generateMedia:
    post:
      description: Generate vocabulary flashcards with audio from a text prompt and
        uploaded images
      operationId: generateMedia
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/GenerateMediaRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
          description: Successfully generated flashcard content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request - invalid input parameters or files
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Payload too large - file size exceeds limits
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error
      summary: Generate flashcard content from text prompt and images
      tags:
      - content-generation
components:
  schemas:
    GenerateTextRequest:
      example:
        targetLang: zh-CN
        maxTerms: 4
        maxMeaningsPerTerm: 6
        withTTS: false
        prompt: HSK2向けに『替・觉得・认为』。各語義ごとに短義＋例文を。
      properties:
        prompt:
          description: Text prompt for generating flashcard content
          example: HSK2向けに『替・觉得・认为』。各語義ごとに短義＋例文を。
          title: prompt
          type: string
        targetLang:
          description: Target language code (ISO 639-1)
          enum:
          - zh-CN
          - zh-TW
          - ja-JP
          - ko-KR
          - en-US
          - es-ES
          - fr-FR
          - de-DE
          example: zh-CN
          title: targetLang
          type: string
        withTTS:
          default: false
          description: Whether to generate text-to-speech audio
          title: withTTS
          type: boolean
        maxTerms:
          default: 10
          description: Maximum number of terms to generate
          maximum: 50
          minimum: 1
          title: maxTerms
          type: integer
        maxMeaningsPerTerm:
          default: 3
          description: Maximum number of meanings per term
          maximum: 10
          minimum: 1
          title: maxMeaningsPerTerm
          type: integer
      required:
      - prompt
      - targetLang
      title: GenerateTextRequest
      type: object
    GenerateMediaRequest:
      properties:
        prompt:
          description: Optional text prompt for generating flashcard content
          example: Generate vocabulary from these images
          type: string
        targetLang:
          description: Target language code (ISO 639-1)
          enum:
          - zh-CN
          - zh-TW
          - ja-JP
          - ko-KR
          - en-US
          - es-ES
          - fr-FR
          - de-DE
          example: zh-CN
          type: string
        withTTS:
          default: false
          description: Whether to generate text-to-speech audio
          type: boolean
        maxTerms:
          default: 10
          description: Maximum number of terms to generate
          maximum: 50
          minimum: 1
          type: integer
        maxMeaningsPerTerm:
          default: 3
          description: Maximum number of meanings per term
          maximum: 10
          minimum: 1
          type: integer
        images:
          description: Array of image files (1 to 10 files supported)
          items:
            format: binary
            type: string
          maxItems: 10
          minItems: 1
          type: array
      required:
      - targetLang
      type: object
    GenerateResponse:
      example:
        items:
        - generated: "### 替\\n- 読み: tì\\n- 品詞: 動詞\\n\\n1) …の代わりにする\\n   - 例文: 他替我参\
            加了会议。\\n   - 訳: 彼は私の代わりに会議に参加しました。"
          term: 替
          audio:
          - ext: mp3
            filename: ti_sense1_norm.mp3
            meaningIndex: 1
            variant: normal
            base64: BASE64_MP3_NORMAL
          - ext: mp3
            filename: ti_sense1_norm.mp3
            meaningIndex: 1
            variant: normal
            base64: BASE64_MP3_NORMAL
        - generated: "### 替\\n- 読み: tì\\n- 品詞: 動詞\\n\\n1) …の代わりにする\\n   - 例文: 他替我参\
            加了会议。\\n   - 訳: 彼は私の代わりに会議に参加しました。"
          term: 替
          audio:
          - ext: mp3
            filename: ti_sense1_norm.mp3
            meaningIndex: 1
            variant: normal
            base64: BASE64_MP3_NORMAL
          - ext: mp3
            filename: ti_sense1_norm.mp3
            meaningIndex: 1
            variant: normal
            base64: BASE64_MP3_NORMAL
      properties:
        items:
          description: Generated flashcard items
          items:
            $ref: "#/components/schemas/FlashcardItem"
          title: items
          type: array
      required:
      - items
      title: GenerateResponse
      type: object
    FlashcardItem:
      example:
        generated: "### 替\\n- 読み: tì\\n- 品詞: 動詞\\n\\n1) …の代わりにする\\n   - 例文: 他替我参加了\
          会议。\\n   - 訳: 彼は私の代わりに会議に参加しました。"
        term: 替
        audio:
        - ext: mp3
          filename: ti_sense1_norm.mp3
          meaningIndex: 1
          variant: normal
          base64: BASE64_MP3_NORMAL
        - ext: mp3
          filename: ti_sense1_norm.mp3
          meaningIndex: 1
          variant: normal
          base64: BASE64_MP3_NORMAL
      properties:
        term:
          description: The vocabulary term
          example: 替
          title: term
          type: string
        generated:
          description: "Generated content in markdown format including pronunciation,\
            \ meanings, and examples"
          example: "### 替\\n- 読み: tì\\n- 品詞: 動詞\\n\\n1) …の代わりにする\\n   - 例文: 他替我参加了\
            会议。\\n   - 訳: 彼は私の代わりに会議に参加しました。"
          title: generated
          type: string
        audio:
          description: Audio files for different meanings (if withTTS is true)
          items:
            $ref: "#/components/schemas/AudioItem"
          title: audio
          type: array
      required:
      - generated
      - term
      title: FlashcardItem
      type: object
    AudioItem:
      example:
        ext: mp3
        filename: ti_sense1_norm.mp3
        meaningIndex: 1
        variant: normal
        base64: BASE64_MP3_NORMAL
      properties:
        meaningIndex:
          description: Index of the meaning this audio corresponds to
          example: 1
          title: meaningIndex
          type: integer
        variant:
          description: Audio variant type
          enum:
          - normal
          - slow
          example: normal
          title: variant
          type: string
        filename:
          description: Generated filename for the audio
          example: ti_sense1_norm.mp3
          title: filename
          type: string
        ext:
          description: File extension
          enum:
          - mp3
          - wav
          example: mp3
          title: ext
          type: string
        base64:
          description: Base64 encoded audio data
          example: BASE64_MP3_NORMAL
          title: base64
          type: string
      required:
      - base64
      - ext
      - filename
      - meaningIndex
      - variant
      title: AudioItem
      type: object
    ErrorResponse:
      example:
        details:
          key: ""
        error: validation_error
        message: The prompt field is required
      properties:
        error:
          description: Error type
          example: validation_error
          title: error
          type: string
        message:
          description: Human-readable error message
          example: The prompt field is required
          title: message
          type: string
        details:
          additionalProperties: true
          description: Additional error details
          title: details
          type: object
      required:
      - error
      - message
      title: ErrorResponse
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http